Implement notifications parsing
