+++ src/com/owncloud/android/lib/resources/notifications/GetRemoteNotificationsOperation.java
@@ -1,8 +1,11 @@
 /*  Nextcloud Android Library is available under MIT license
  *   Copyright (C) 2017 Andy Scherzinger
+ *   Copyright (C) 2017 Mario Danic
  *   Copyright (C) 2017 Nextcloud GmbH
  *
  *   @author Andy Scherzinger
+ *   @author Mario Danic
+ *
  *
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
  *   of this software and associated documentation files (the "Software"), to deal
@@ -27,19 +30,23 @@
 
 package com.owncloud.android.lib.resources.notifications;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import com.google.gson.reflect.TypeToken;
 import com.owncloud.android.lib.common.OwnCloudClient;
 import com.owncloud.android.lib.common.operations.RemoteOperation;
 import com.owncloud.android.lib.common.operations.RemoteOperationResult;
 import com.owncloud.android.lib.common.utils.Log_OC;
+import com.owncloud.android.lib.resources.notifications.models.Notification;
 
 import org.apache.commons.httpclient.HttpStatus;
 import org.apache.commons.httpclient.methods.GetMethod;
 import org.json.JSONException;
-import org.json.JSONObject;
 
-import java.text.ParseException;
+import java.lang.reflect.Type;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -51,50 +58,14 @@ public class GetRemoteNotificationsOperation extends RemoteOperation {
 
     // OCS Route
     private static final String OCS_ROUTE_LIST =
-            "/ocs/v2.php/apps/notifications/api/v2/notifications";
+            "/ocs/v2.php/apps/notifications/api/v2/notifications?format=json";
 
     private static final String TAG = GetRemoteNotificationsOperation.class.getSimpleName();
 
     // JSON Node names
     private static final String NODE_OCS = "ocs";
 
-    private static final String NODE_OCS_META = "meta";
-    private static final String NODE_OCS_META_STATUS = "status";
-    private static final String NODE_OCS_META_STATUS_CODE = "statuscode";
-    private static final String NODE_OCS_META_STATUS_MESSAGE = "message";
-
     private static final String NODE_DATA = "data";
-    private static final String NODE_NOTIFICATION_ID = "notification_id";
-    private static final String NODE_APP = "app";
-    private static final String NODE_USER = "user";
-    private static final String NODE_DATE_TIME = "datetime";
-    private static final String NODE_OBJECT_TYPE = "object_type";
-    private static final String NODE_OBJECT_ID = "object_id";
-    private static final String NODE_SUBJECT = "subject";
-
-    private static final String NODE_SUBJECT_RICH = "subjectRich";
-    private static final String NODE_SUBJECT_RICH_PARAMS = "subjectRichParameters";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE = "share";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_TYPE = "type";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_TYPE_ID = "id";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_TYPE_NAME = "name";
-
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_USER = "user";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_USER_TYPE = "type";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_USER_TYPE_ID = "id";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_USER_TYPE_NAME = "name";
-    private static final String NODE_SUBJECT_RICH_PARAMS_SHARE_USER_TYPE_SERVER = "server";
-
-    private static final String NODE_MESSAGE = "message";
-    private static final String NODE_MESSAGE_RICH = "messageRich";
-    private static final String NODE_MESSAGE_RICH_PARAMS = "messageRichParameters";
-    private static final String NODE_LINK = "link";
-    private static final String NODE_ICON = "icon";
-    private static final String NODE_ACTIONS = "actions";
-    private static final String NODE_ACTIONS_LABEL = "label";
-    private static final String NODE_ACTIONS_LABEL_LINK = "link";
-    private static final String NODE_ACTIONS_LABEL_TYPE = "type";
-    private static final String NODE_ACTIONS_LABEL_PRIMARY = "primary";
 
     /**
      * Date pattern according to h
@@ -114,6 +85,7 @@ public class GetRemoteNotificationsOperation extends RemoteOperation {
         RemoteOperationResult result = null;
         int status = -1;
         GetMethod get = null;
+        List<Notification> notifications;
         String url = client.getBaseUri() + OCS_ROUTE_LIST;
 
         // get the notifications
@@ -122,13 +94,22 @@ public class GetRemoteNotificationsOperation extends RemoteOperation {
             get.addRequestHeader(OCS_API_HEADER, OCS_API_HEADER_VALUE);
 
             status = client.executeMethod(get);
+            String response = get.getResponseBodyAsString();
 
             if (isSuccess(status)) {
-                String response = get.getResponseBodyAsString();
                 Log_OC.d(TAG, "Successful response: " + response);
 
                 // Parse the response
-                Notification notification = parseResult(response);
+                notifications = parseResult(response);
+                result.setNotificationData(notifications);
+            } else {
+                result = new RemoteOperationResult(false, status, get.getResponseHeaders());
+                Log_OC.e(TAG, "Failed response while getting user notifications ");
+                if (response != null) {
+                    Log_OC.e(TAG, "*** status code: " + status + " ; response message: " + response);
+                } else {
+                    Log_OC.e(TAG, "*** status code: " + status);
+                }
             }
         } catch (Exception e) {
             result = new RemoteOperationResult(e);
@@ -142,66 +123,16 @@ public class GetRemoteNotificationsOperation extends RemoteOperation {
         return result;
     }
 
-    private Notification parseResult(String response) throws JSONException {
-        JSONObject respJSON = new JSONObject(response);
-        JSONObject respOCS = respJSON.getJSONObject(NODE_OCS);
-        JSONObject respData = respOCS.getJSONObject(NODE_DATA);
+    private List<Notification> parseResult(String response) throws JSONException {
+        JsonParser jsonParser = new JsonParser();
+        JsonObject jo = (JsonObject)jsonParser.parse(response);
+        JsonArray jsonDataArray = jo.getAsJsonObject(NODE_OCS).getAsJsonArray(NODE_DATA);
 
-        Notification notification = new Notification();
+        Gson gson = new Gson();
+        Type listType = new TypeToken<List<Notification>>(){}.getType();
+        List<Notification> notifications = gson.fromJson(jsonDataArray, listType);
 
-        notification.setNotificationId(respData.getInt(NODE_NOTIFICATION_ID));
-        notification.setApp(respData.getString(NODE_APP));
-        notification.setUser(respData.getString(NODE_USER));
-        try {
-            notification.setDatetime(SDF.parse(respData.getString(NODE_DATE_TIME)));
-        } catch (ParseException e) {
-            Log_OC.e(TAG, "Error parsing date "+ respData.getString(NODE_DATE_TIME), e);
-        }
-        notification.setObject_type(respData.getString(NODE_OBJECT_TYPE));
-        notification.setObject_id(respData.getString(NODE_OBJECT_ID));
-        notification.setSubject(respData.getString(NODE_SUBJECT));
-
-        // optional parameters
-        //v2+ only
-        if (respData.has(NODE_SUBJECT_RICH)) {
-            notification.setIcon(respData.getString(NODE_SUBJECT_RICH));
-        }
-
-        //v2+ only
-        if (respData.has(NODE_SUBJECT_RICH_PARAMS)) {
-            // Todo loop through array
-        }
-
-        if (respData.has(NODE_MESSAGE)) {
-            notification.setMessage(respData.getString(NODE_MESSAGE));
-        }
-
-        //v2+ only
-        if (respData.has(NODE_MESSAGE_RICH)) {
-            notification.setMessage(respData.getString(NODE_MESSAGE_RICH));
-        }
-
-        //v2+ only
-        if (respData.has(NODE_MESSAGE_RICH_PARAMS)) {
-            // Todo loop through array
-            List<RichObject> messageRichObjects = new ArrayList<RichObject>();
-        }
-
-        if (respData.has(NODE_LINK)) {
-            notification.setLink(respData.getString(NODE_LINK));
-        }
-
-        //v2+ only
-        if (respData.has(NODE_ICON)) {
-            notification.setIcon(respData.getString(NODE_ICON));
-        }
-
-        if (respData.has(NODE_ACTIONS)) {
-            // Todo loop through array
-            List<Action> actions = new ArrayList<Action>();
-        }
-
-        return notification;
+        return notifications;
     }
 
     private boolean isSuccess(int status) {
