+++ src/com/owncloud/android/lib/resources/notifications/models/Action.java
@@ -25,59 +25,39 @@
  *
  */
 
-package com.owncloud.android.lib.resources.notifications;
+package com.owncloud.android.lib.resources.notifications.models;
 
 /**
- * Rich object data model providing rich data to be used within rich content, e.g.
- * in {@link Notification}s.
+ * Action data model.
  */
-public class RichObject {
+public class Action {
+
     /**
-     * A unique identifier for the object type.
+     * Translated short label of the action/button that should be presented to the user.
+     */
+    public String label;
+    /**
+     * A link that should be followed when the action is performed/clicked.
+     */
+    public String link;
+    /**
+     * HTTP method that should be used for the request against the link: GET, POST, DELETE.
      */
     public String type;
 
     /**
-     * A short identifier of the object on the server (int or string).
+     * If the action is the primary action for the notification or not.
      */
-    public String id;
+    public boolean primary;
 
-    /**
-     * A name which should be used in the visual representation.
-     */
-    public String name;
-
-    public RichObject() {
+    public Action() {
 
     }
 
-    public RichObject(String type, String id, String name) {
+    public Action(String label, String link, String type, boolean primary) {
+        this.label = label;
+        this.link = link;
         this.type = type;
-        this.id = id;
-        this.name = name;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
+        this.primary = primary;
     }
 }
